add dendrite query metadata
